pull the number of flows from the query service

+++ rest-server/src/main/java/org/opennms/horizon/server/model/flows/FlowSummary.java
@@ -6,5 +6,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class FlowSummary {
-    int numFlows;
+    long numFlows;
 }
